refactor(workers): split EmbodiedFSDPActor into its own module
Move EmbodiedFSDPActor and its four nested-dict helpers
(process_nested_dict_for_adv, process_nested_dict_for_train,
trim_nested_tensor_time_dim, flatten_nested_tensor_time_batch) from
fsdp_actor_worker.py into a new embodied_fsdp_actor_worker.py.
fsdp_actor_worker.py now only contains FSDPActor (reasoning/LLM RL)
and compute_rollout_train_kl, and updates the 8 downstream modules
that imported EmbodiedFSDPActor from the old module path.

fsdp_actor_worker.py goes from 1731 to 976 lines; the new file is 792
lines. Verified with ruff, an import smoke test of all subclasses, and a
libero_goal_ppo_openpi_pi05 e2e run (one full rollout->train iteration,
47 tensorboard scalars including train/actor/{grad_norm,total_loss,
policy_loss}).
